chore: initial di pattern
