+++ src/modules/room/handlers/room-leave.handler.ts
@@ -1,0 +1,74 @@
+import ChannelManager from '@/lib/amqp-manager/channel-manager';
+import { RedisClient } from '@/lib/redis';
+import { Session } from '@/types/session.types';
+import { SocketAckHandler } from '@/types/socket.types';
+import { getNspRoomId } from '@/util/helpers';
+import { getLogger } from '@/util/logger';
+import { WebSocket } from 'uWebSockets.js';
+import { leaveRoom } from '@/modules/room/room.service';
+import { pushRoomLeaveMetrics } from '@/modules/metrics/metrics.service';
+import { enqueueWebhookEvent } from '@/modules/webhook/webhook.service';
+import { WebhookEvent } from '@/types/webhook.types';
+import { formatErrorResponse, formatPresenceSubscription } from '@/util/format';
+import { ClientEvent } from '@/types/event.types';
+import { SubscriptionType } from '@/types/subscription.types';
+import { removeActiveMember } from '@/modules/presence/presence.service';
+import { unbindAllSubscriptions } from '@/modules/subscription/subscription.service';
+import { KeyNamespace } from '@/types/state.types';
+import { Logger } from 'winston';
+
+const logger = getLogger(ClientEvent.ROOM_LEAVE);
+
+export async function handler(
+  logger: Logger, // TODO: REMOVE
+  redisClient: RedisClient,
+  socket: WebSocket<Session>,
+  data: any,
+  res: SocketAckHandler
+): Promise<void> {
+  const session = socket.getUserData();
+
+  const { roomId } = data;
+  const { uid, connectionId, user, clientId } = session;
+
+  try {
+    const nspRoomId = getNspRoomId(session.appPid, roomId);
+    const nspRoomRoutingKey = ChannelManager.getRoutingKey(nspRoomId);
+    const presenceSubsciption = formatPresenceSubscription(nspRoomId, SubscriptionType.LEAVE);
+    const timestamp = new Date().toISOString();
+
+    const presenceLeaveMessage = {
+      clientId,
+      event: SubscriptionType.LEAVE,
+      user,
+      timestamp
+    };
+
+    const webhookdata = {
+      roomId
+    };
+
+    await Promise.all([
+      leaveRoom(redisClient, session, nspRoomId, socket),
+      leaveRoom(redisClient, session, nspRoomRoutingKey, socket),
+      removeActiveMember(uid, nspRoomId, presenceSubsciption, session, presenceLeaveMessage),
+      unbindAllSubscriptions(
+        redisClient,
+        connectionId,
+        nspRoomId,
+        KeyNamespace.SUBSCRIPTIONS,
+        socket
+      ),
+      unbindAllSubscriptions(redisClient, connectionId, nspRoomId, KeyNamespace.PRESENCE, socket),
+      unbindAllSubscriptions(redisClient, connectionId, nspRoomId, KeyNamespace.METRICS, socket),
+      pushRoomLeaveMetrics(uid, nspRoomId, session),
+      enqueueWebhookEvent(WebhookEvent.ROOM_LEAVE, webhookdata, session)
+    ]);
+
+    res(nspRoomId);
+  } catch (err: any) {
+    logger.error(`Failed to leave room "${roomId}"`, { err, roomId, session });
+
+    res(null, formatErrorResponse(err));
+  }
+}
